+++ src/main/java/com/bairei/ormapp/services/implementation/BandServiceImpl.java
@@ -1,8 +1,10 @@
 package com.bairei.ormapp.services.implementation;
 
+import com.bairei.ormapp.models.Album;
 import com.bairei.ormapp.models.Band;
 import com.bairei.ormapp.models.Event;
 import com.bairei.ormapp.repositories.BandRepository;
+import com.bairei.ormapp.services.AlbumService;
 import com.bairei.ormapp.services.EventService;
 import com.bairei.ormapp.services.BandService;
 import lombok.extern.slf4j.Slf4j;
@@ -19,10 +21,12 @@ public class BandServiceImpl implements BandService {
 
     private final BandRepository bandRepository;
     private final EventService eventService;
+    private final AlbumService albumService;
 
-    public BandServiceImpl(BandRepository bandRepository, EventService eventService) {
+    public BandServiceImpl(BandRepository bandRepository, EventService eventService, AlbumService albumService) {
         this.bandRepository = bandRepository;
         this.eventService = eventService;
+        this.albumService = albumService;
     }
 
     @Override
@@ -50,22 +54,27 @@ public class BandServiceImpl implements BandService {
     @Transactional
     public void deleteById(Long aLong) {
         Band band = bandRepository.findById(aLong);
-        band.setGenre(null);
-        band.setMembers(new HashSet<>());
-        bandRepository.save(band);
-        for (Event event: eventService.findEventsByBandsIncludingBandNameEqualsIgnoreCase(band.getName())){
-            if(event.getBandSet().contains(band)) {
-                Set<Band> bands = event.getBandSet();
-                bands.remove(band);
-                event.setBandSet(bands);
-                if(event.getBandSet().size() < 1){
-                    eventService.deleteById(event.getId());
-                } else {
-                    eventService.save(event);
+        if (band != null) {
+            band.setGenre(null);
+            band.setMembers(new HashSet<>());
+            bandRepository.save(band);
+            for (Event event : eventService.findEventsByBandsIncludingBandNameEqualsIgnoreCase(band.getName())) {
+                if (event.getBandSet().contains(band)) {
+                    Set<Band> bands = event.getBandSet();
+                    bands.remove(band);
+                    event.setBandSet(bands);
+                    if (event.getBandSet().size() < 1) {
+                        eventService.deleteById(event.getId());
+                    } else {
+                        eventService.save(event);
+                    }
                 }
             }
+            for (Album a: band.getAlbums()){
+                albumService.deleteById(a.getId());
+            }
+            bandRepository.deleteById(aLong);
         }
-        bandRepository.deleteById(aLong);
     }
 
     @Override
